+++ eggs/Naaya/CHANGELOG.rst
@@ -4,6 +4,7 @@
 * fix update script to remove duplicate images [dumitval]
 * remove inline styles [moregale]
 * ``naaya:simpleView`` directive [moregale]
+* ``naaya:rstkMethod`` directive [moregale]
 
 2.12.53 (2012-01-24)
 --------------------